refactor(experience): shared tag overflow like Projects; revert hover motion
- Extract the Projects single-line "+N tooltip" tag logic into a shared
  TagRow component and use it on both pages, so the Experience skills cut off
  and tooltip identically instead of wrapping. Aligned to the max-w-2xl column.
- Revert the row hover back to a plain CSS background transition (the motion
  attempt wasn't what was wanted).
