+++ src/pages/Projects.tsx
@@ -6,13 +6,7 @@
  * Refer to LICENSE for details or contact yanis.sebastian.zuercher@gmail.com for permissions.
  */
 
-import {
-  useState,
-  useMemo,
-  useRef,
-  useLayoutEffect,
-  type ReactNode,
-} from "react";
+import { useState, useMemo, type ReactNode } from "react";
 import {
   ArrowUpRight,
   SortAsc,
@@ -48,8 +42,8 @@ import {
 } from "@/components/ui/tooltip";
 import { Skeleton } from "@/components/ui/skeleton";
 import { Card } from "@/components/ui/card";
-import { Badge } from "@/components/ui/badge";
 import { Separator } from "@/components/ui/separator";
+import { TagRow } from "@/components/ui/custom/tag-row";
 import { cn } from "@/lib/utils";
 import { PROJECTS, type ProjectMeta } from "@/config/projects";
 
@@ -146,109 +140,6 @@ const ProjectImage = ({ project, t }: { project: Project; t: Translation }) => {
   );
 };
 
-const TAG_GAP = 6;
-
-const ProjectTechnologies = ({
-  technologies,
-}: {
-  technologies: string[];
-}) => {
-  const containerRef = useRef<HTMLDivElement>(null);
-  const [visibleCount, setVisibleCount] = useState(technologies.length);
-
-  useLayoutEffect(() => {
-    const container = containerRef.current;
-    if (!container) return;
-    let disposed = false;
-
-    const compute = () => {
-      if (disposed) return;
-      const available = Math.floor(container.clientWidth);
-      if (!available) return;
-
-      const widths = Array.from(
-        container.querySelectorAll<HTMLElement>("[data-measure='tag']"),
-      ).map((el) => el.offsetWidth);
-      const moreWidth =
-        container.querySelector<HTMLElement>("[data-measure='more']")
-          ?.offsetWidth ?? 0;
-
-      let used = 0;
-      let count = 0;
-      for (let i = 0; i < widths.length; i++) {
-        const width = widths[i] + (i > 0 ? TAG_GAP : 0);
-        if (used + width > available) break;
-        used += width;
-        count++;
-      }
-
-      if (count < widths.length) {
-        while (count > 0 && used + TAG_GAP + moreWidth > available) {
-          used -= widths[count - 1] + (count - 1 > 0 ? TAG_GAP : 0);
-          count--;
-        }
-      }
-
-      setVisibleCount(count);
-    };
-
-    const observer = new ResizeObserver(compute);
-    observer.observe(container);
-    compute();
-    document.fonts?.ready.then(compute);
-
-    return () => {
-      disposed = true;
-      observer.disconnect();
-    };
-  }, [technologies]);
-
-  const hiddenCount = technologies.length - visibleCount;
-
-  return (
-    <div ref={containerRef} className="relative flex gap-1.5 overflow-hidden">
-      <div
-        aria-hidden
-        className="pointer-events-none invisible absolute left-0 top-0 flex gap-1.5"
-      >
-        {technologies.map((technology) => (
-          <Badge
-            key={technology}
-            data-measure="tag"
-            variant="secondary"
-            className="font-normal"
-          >
-            {technology}
-          </Badge>
-        ))}
-        <Badge data-measure="more" variant="secondary" className="font-normal">
-          +{technologies.length}
-        </Badge>
-      </div>
-
-      {technologies.slice(0, visibleCount).map((technology) => (
-        <Badge key={technology} variant="secondary" className="font-normal">
-          {technology}
-        </Badge>
-      ))}
-      {hiddenCount > 0 && (
-        <Tooltip>
-          <TooltipTrigger
-            render={
-              <Badge variant="secondary" className="font-normal cursor-default">
-                +{hiddenCount}
-              </Badge>
-            }
-          />
-          <TooltipContent className="max-w-[220px] text-center">
-            {technologies.slice(visibleCount).join(", ")}
-          </TooltipContent>
-        </Tooltip>
-      )}
-    </div>
-  );
-};
-
 const ProjectActions = ({
   project,
   t,
@@ -328,7 +219,7 @@ const ProjectBody = ({ project, t }: { project: Project; t: Translation }) => (
       <RichText text={project.description} />
     </p>
     {project.technologies.length > 0 && (
-      <ProjectTechnologies technologies={project.technologies} />
+      <TagRow tags={project.technologies} />
     )}
     <ProjectActions project={project} t={t} />
   </div>
